Add files via upload
This is the data insight that have been created using Power BI Desktop tool. The sample dataset used for the project is the sales record of a superstore, which is stored in an MS Excel workbook. It is structured dataset with fields like Order-ID, Customer-ID, Customer Name, State, Country, Ship Date, Shipping Order, etc.